adding Symbole table for label handling

+++ src/machine/EnvRegisters.cpp
@@ -8,8 +8,8 @@ EnvRegisters::EnvRegisters()
 
     GB = std::make_unique<Register>("GB",Value(0));
     LB = std::make_unique<Register>("LB",Value(0));
-    SP = std::make_unique<Register>("SP",Value(0));
-    PC = std::make_unique<Register>("PC",Value(0));
+    SP = std::make_unique<Register>("SP",Value(static_cast<uint32_t>(0)));
+    PC = std::make_unique<Register>("PC",Value(static_cast<uint32_t>(0)));
     
     for(int i = 0 ; i < NUMBER_MAX_GENPORPSE_REG ; i++)
     {
